feat(about): remove goodreads widget
Feels like the goodreads widget become like a glitch. First, they show
the placeholder of the recent books in the past, and then they fetch
the new recent books to show the updated recent books.

I feel like this whole process is like a glitch, so let's remove it and
make about page simpler.