chore(posts): add fix skeletyl keyboard not useable from booting
In one of my laptop, skeletyl keyboard is not functioning after booting
the laptop, the cold boot from shutdown state and not from reboot state,
and i need to re-plug the keyboard to be able to use it.

In this changes, add info about how to fix those problem.

And also add info about flashing both side just to be safe.