chore(posts): remove template.Must() and handle error manually
In the posts "golang simple query composer with text template", i use
`template.Must()` to make sure that the template parsing successful. The
problem is that, `template.Must()` use `panic()` which can cause the API
server crash if we failed to parse the template and that could be
dangerous. So i changed that to use `template.New().Parse()` directly
without `template.Must()` and handle the error manually with
`if err != nil`.