feat(posts): add competing on the worst experience
This is my thought after watching Dr. K's youtube video about "why 'the
grind' isn't for everyone".

This blog post is more of the thought about one of his statement, that
is "we live in a society that glorifies exhaustion at work", rather than
'the grind' itself.

The point is that, i think people is not only glorifies exhaustion
or any bad experience, they competing to see who has the most exhaustion
or the worst experience, and that's kind of concerning.