feat(posts): add post about artificial as replacement point of view
This post is another point of view when we are talking about
something that have "artificial" in the name, like "artificial leg",
"artificial arm", or "artificial flowers" which act as a replacement for
the real thing.

With that in mind, what about "artificial intelligence"? What are we
trying to replace with "artificial intelligence"? That's an interesting
question, right?