feat(posts): add square bracket as dereference operator in C
This post is what i learned about square bracket in C programming
language, which is actually another dereference operator. So we are not
only have asterisk (*) as a dereference operator, we can also use square
bracket as dereference operator.

So writing this post as notes for my future self or someone else who
want to learn about C programming language.